Rename beman.example to beman.exemplar
Calling this project exemplar is more accurate and mitigates some of the
confusion created by having an "examples" folder within an "example"
project.

Once this chnage merges, I intend to rename the repository to exemplar.
