+++ .ci/docker/ubuntu.Dockerfile
@@ -30,5 +30,5 @@ ENV CC="$cc" CXX="$cxx" CMAKE_GENERATOR="Ninja" CMAKE_EXPORT_COMPILE_COMMANDS=on
 RUN ls -lR src
 RUN cmake -B build -S . "$cmake_args"
 RUN cmake --build build --verbose
-RUN cmake --install build --prefix /opt/beman.example
-RUN find /opt/beman.example -type f
+RUN cmake --install build --prefix /opt/beman.exemplar
+RUN find /opt/beman.exemplar -type f
