+++ examples/identity_direct_usage.cpp
@@ -1,10 +1,10 @@
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include <beman/example/identity.hpp>
+#include <beman/exemplar/identity.hpp>
 
 #include <iostream>
 
 int main() {
-    std::cout << beman::example::identity()(2024) << '\n';
+    std::cout << beman::exemplar::identity()(2024) << '\n';
     return 0;
 }
