+++ include/beman/exemplar/identity.hpp
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#ifndef BEMAN_EXAMPLE_IDENTITY_HPP
-#define BEMAN_EXAMPLE_IDENTITY_HPP
+#ifndef BEMAN_EXEMPLAR_IDENTITY_HPP
+#define BEMAN_EXEMPLAR_IDENTITY_HPP
 
 // C++ Standard Library: std::identity equivalent.
 // See https://eel.is/c++draft/func.identity:
@@ -22,7 +22,7 @@
 
 #include <utility> // std::forward
 
-namespace beman::example {
+namespace beman::exemplar {
 
 struct __is_transparent; // not defined
 
@@ -43,6 +43,6 @@ struct identity
     using is_transparent = __is_transparent;
 };
 
-} // namespace beman::example
+} // namespace beman::exemplar
 
-#endif // BEMAN_EXAMPLE_IDENTITY_HPP
+#endif // BEMAN_EXEMPLAR_IDENTITY_HPP
