+++ src/beman/exemplar/identity.cpp
@@ -1,3 +1,3 @@
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include <beman/example/identity.hpp>
+#include <beman/exemplar/identity.hpp>
